requirements.txt
flask
flask-cors
newsapi-python